Move opensea fetch from serverless to pure github action
