+++ package.json
@@ -5,6 +5,7 @@
     "web": "yarn workspace @dopewars/web",
     "web:dev": "yarn workspace @dopewars/web dev",
     "web:test": "yarn workspace @dopewars/web test",
+    "web:refresh-opensea-sales-cache": "yarn workspace @dopewars/web refresh-opensea-sales-cache",
     "contracts": "yarn workspace @dopewars/contracts",
     "lint": "eslint 'packages/web/**/*.ts' --fix",
     "format": "prettier --write 'packages/web/**/*.{ts(x)?,sol,md,css}' '!**/typechain/**'"
