+++ packages/web/package.json
@@ -9,7 +9,8 @@
     "gen": "graphql-codegen --config codegen.yml",
     "test": "jest",
     "test-watch": "jest --watch",
-    "dump-opensea": "ts-node --script-mode script/dump-opensea-json.ts"
+    "dump-opensea": "ts-node --script-mode script/dump-opensea-json.ts",
+    "refresh-opensea-sales-cache": "ts-node --script-mode script/refresh-opensea-sales-cache.ts"
   },
   "dependencies": {
     "@apollo/client": "^3.4.11",
@@ -23,6 +24,7 @@
     "aws-sdk": "^2.999.0",
     "classnames": "^2.3.1",
     "dope-metrics": "^1.0.4",
+    "dotenv": "^10.0.0",
     "framer-motion": "^4",
     "gear-rarity": "^1.0.5",
     "graphql": "^15.5.3",
